+++ CHANGELOG.rst
@@ -2,6 +2,18 @@
 Changelog
 =========
 
+-----------------------
+v0.2 (work in progress)
+-----------------------
+
+- Making date parser for HPO annotation files more robust.
+  It works now for positive and negative associations.
+- Small bug fix in HPO OBO parser.
+- Adding ``ontolib-cli`` package that allows score distribution precomputation from the command line.
+- Removed some dead code.
+- Added various tests, minor internal refactoring.
+- Moved ``OntologyTerms`` into ``ontology.algo`` package.
+
 ----
 v0.1
 ----